feat: remove .types files
