--- src/components/Toast/Toast.types.ts
@@ -1,14 +1,0 @@
-export enum ToastType {
-  INFO = 'info',
-  WARN = 'warn',
-  ERROR = 'error'
-}
-
-export type Props = {
-  type?: ToastType
-  title: string | JSX.Element
-  body: string | JSX.Element
-  closable?: boolean
-  timeout?: number
-  onClose?: () => void
-}
