+++ src/components/Toasts/Toasts.tsx
@@ -1,9 +1,18 @@
 import * as React from 'react'
-
-import { Props } from './Toasts.types'
 import './Toasts.css'
 
-export default class Toasts extends React.PureComponent<Props> {
+export type ToastPosition =
+  | 'bottom left'
+  | 'bottom right'
+  | 'top left'
+  | 'top right'
+
+export type ToastsProps = {
+  position?: ToastPosition
+  children: React.ReactNode
+}
+
+export default class Toasts extends React.PureComponent<ToastsProps> {
   render() {
     const { position = 'top right', children } = this.props
